+++ src/main/java/fr/raksrinana/fallingtree/tree/builder/AbortSearchException.java
@@ -1,9 +1,7 @@
 package fr.raksrinana.fallingtree.tree.builder;
 
-import net.minecraft.block.Block;
-
 public class AbortSearchException extends RuntimeException{
-	public AbortSearchException(Block block){
-		super("Found block " + block + " that isn't whitelisted");
+	public AbortSearchException(String reason){
+		super(reason);
 	}
 }